refactor: extract pure toMermaid builder for app + CLI reuse
